refactor: move saving to FileManager
